show error if no cameras available issue #54
